no_engine: overhaul projects adding `no_engine_sys`
## Summary

This change adds the `no_engine_sys` which bundles up a bunch of
no_engine types. It goes on to provide a `log_sys` and store contexts
directly in systems.

This effort is focused on making this package easier to use and getting
a proper logging implementation up and running.